Fixed MatchOddsService and now it saves the entities correctly. Also added unique constraint to match_id and specifier and made a more user friendly error for unique constaint violation.

+++ src/main/java/com/fotistsiskakis/betstrategist/models/requests/CreateMatchOddsRequest.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -14,5 +15,5 @@ public class CreateMatchOddsRequest {
     @NotNull(message = "specifier cannot be null")
     private String specifier;
     @NotNull(message = "odd cannot be null")
-    private Long odd;
+    private BigDecimal odd;
 }
